requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
requests==2.31.0
uvicorn==0.23.2
gunicorn==21.2.0